+++ docs/releases.md
@@ -2,6 +2,12 @@
 
 ## Release next
 
+### Fix #52: Duplicate command fails with 504 timeout
+A new option enable the possibility of calling directly the immich-server once it's port is published.
+
+`-api URL` URL of the Immich api endpoint (http://container_ip:3301)<br>
+
+
 ### Fix #50:  Duplicate detection fails when timezone of both images differs
 Imported duplicated images with same name, but different timezone wasn't seen as duplicates.
 The `-ignore-tz-errors=true` compares the time on date, and minute and ignores the hour of capture.